MAINT: Suggested changes after review. Moved methods of MCOModel to MCO (tests updated), `MockOptimizer` moved to tests folder, naming changes, log strings concatenated, typos fixed.
