+++ itwm_example/mco/optimizers/optimizers.py
@@ -4,13 +4,7 @@ from functools import partial
 import numpy as np
 from scipy import optimize as scipy_optimize
 
-from traits.api import (
-    Interface,
-    HasStrictTraits,
-    provides,
-    List,
-    Instance,
-)
+from traits.api import Interface, HasStrictTraits, provides, List, Instance
 
 from force_bdss.api import BaseMCOParameter
 from force_bdss.mco.i_evaluator import IEvaluator
@@ -53,9 +47,7 @@ class WeightedOptimizer(HasStrictTraits):
 
     def _score(self, point, weights):
 
-        score = np.dot(
-            weights, self.single_point_evaluator.evaluate(point)
-        )
+        score = np.dot(weights, self.single_point_evaluator.evaluate(point))
 
         log.info("Weighted score: {}".format(score))
 
@@ -67,25 +59,17 @@ class WeightedOptimizer(HasStrictTraits):
 
         weighted_score_func = partial(self._score, weights=weights)
 
-        log.info("Running optimisation.")
-        log.info("Initial point: {}".format(initial_point))
-        log.info("Constraints: {}".format(constraints))
+        log.info(
+            "Running optimisation."
+            + "Initial point: {}".format(initial_point)
+            + "Constraints: {}".format(constraints)
+        )
         optimal_point = opt(weighted_score_func, initial_point, constraints)
         optimal_kpis = self.single_point_evaluator.evaluate(optimal_point)
-        log.info("Optimal point : {}".format(optimal_point))
-        log.info("KPIs at optimal point : {}".format(optimal_kpis))
+
+        log.info(
+            "Optimal point : {}".format(optimal_point)
+            + "KPIs at optimal point : {}".format(optimal_kpis)
+        )
 
         return optimal_point, optimal_kpis
-
-
-@provides(IOptimizer)
-class MockOptimizer:
-    def __init__(self, eval, param, **kwargs):
-        self.dimension = 2
-        self.margins = np.array([10.0 for _ in range(self.dimension)])
-        self.min_values = np.array([i for i in range(self.dimension)])
-
-        self.scaling_values = np.array([0.1] * self.dimension)
-
-    def optimize(self, weights):
-        return 0, self.min_values + weights * self.margins
